Hibernate-create/delete/getAll methods. Update is next.

+++ src/main/java/com/addressbook/dto/AbstractContactDTO.java
@@ -10,7 +10,7 @@ public abstract class AbstractContactDTO {
     protected String company;
     protected String phoneNumber;
 
-    public Integer getId() {
+    public long getId() {
         return id;
     }
